js/src/xpconnect/tests/mochitest/test_bug429070.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=429070
-->
<head>
  <title>Test for Bug 429070</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=429070">Mozilla Bug 429070</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 429070 **/
try {
    var a = Components.interfaces;
    ok(false, "Components.interfaces is accessible from content");
} catch(e) {
    ok(true, "Components.interfaces is not exposed to content");
}

try {
    var a = Components.interfacesByID;
    ok(false, "Components.interfacesById is accessible from content");
} catch(e) {
    ok(true, "Components.interfacesById is not exposed to content");
}

try {
    var a = Components.classes;
    ok(false, "Components.classes is accessible from content");
} catch(e) {
    ok(true, "Components.classes is not exposed to content");
}

try {
    var a = Components.classesByID;
    ok(false, "Components.classesByID is accessible from content");
} catch(e) {
    ok(true, "Components.classesByID is not exposed to content");
}

try {
    var a = Components.stack;
    ok(false, "Components.stack is accessible from content");
} catch(e) {
    ok(true, "Components.stack is not exposed to content");
}

try {
    var a = Components.results;
    ok(false, "Components.results is accessible from content");
} catch(e) {
    ok(true, "Components.results is not exposed to content");
}

try {
    var a = Components.manager;
    ok(false, "Components.manager is accessible from content");
} catch(e) {
    ok(true, "Components.manager is not exposed to content");
}

try {
    var a = Components.utils;
    ok(false, "Components.utils is accessible from content");
} catch(e) {
    ok(true, "Components.utils is not exposed to content");
}

try {
    var a = Components.ID;
    ok(false, "Components.ID is accessible from content");
} catch(e) {
    ok(true, "Components.ID is not exposed to content");
}

try {
    var a = Components.Exception;
    ok(false, "Components.Exception is accessible from content");
} catch(e) {
    ok(true, "Components.Exception is not exposed to content");
}

try {
    var a = Components.Constructor;
    ok(false, "Components.Constructor is accessible from content");
} catch(e) {
    ok(true, "Components.Constructor is not exposed to content");
}

</script>
</pre>
</body>
</html>
